tests: add initial testing farm integration
This adds initial testing farm integration via a github action.
To run jobs on testing farm a token is required that is stored
as a repository secret. For security reasons repository secrets
are not visible accross forks [0].

There are multiple ways to work around this limiation, this commit
goes with the suggestion from [1], i.e.: the workflow is run within
the `pull_request_target` trigger which has access to secrets.

This means the (potentially untrusted) branch is only checked out
if the person triggering the workflow has already write access to
the repository (we could make this restriction strong but it seems
a reasonable permisson level). In practise the workflow will fail
for outside contributions but a re-trigger from anyone in the term
should be enough to get it tested inside the testing farm.

[0] https://securitylab.github.com/research/github-actions-preventing-pwn-requests/
[1] https://michaelheap.com/access-secrets-from-forks/