github: update `compose:` value for testing farm GH action
In order to use nested virtualizaton we need to use the
`Red Hat Ranch` from testing farm. The testing farm token is now
updated in the github secrets. However the default value of the
`sclorg/testing-farm-as-github-action@v1` action is to use
`compose: Fedora-latest`. According to:
https://docs.testing-farm.io/Testing%20Farm/0.1/test-environment.html#_composes
this is not available in the RH ranch so we need to use something
else.

I picked `Fedora-39` for now but happy about alternative ideas
(`CentOS-Stream-9` might also be an option?).