test: fix race in wait_ssh_ready() and add test
The wait_ssh_ready() had a subtle bug - when ssh was starting
up and not yet producing the expected output (e.g. just empty
output) the loop that sleeps would not wait but just loop very
quickly which could lead to a confuging:
```
ConnectionRefusedError: cannot connect to port 39973 after 600s
```
even though clearly 600s had not passed.

This commit fixes this (subtle) bug and adds a test that will
generate an unexpected output and ensures there that the expected
number of sleeps happend.