test: skip qemu test when running on non-x86_64-v3 CPUs
Building ELN requires a CPU that is capable of runing `x86_64-v3`
instructions [0]. Most CPUs after 2015 do support this. However
it happend that the test got scheduled inside a lab that had
older CPUs. In this case skip the qemu test. This is okay because
the test is also run in GH runners that are more current.

To skip the `avx2` cpu feature is used as a proxy instead of
testing all of them. This should be a reasonable approximation
but if we run into issues it can always be make more accurate.

[0] https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels