test: refactor monolithic `test_smoke` into multiple tests
The existing test_smoke was becoming a bit too big and especially
the fact that part have to be skipped depending on the environment
is really not nice.

This commit refactors it to be more module. For this it uses the
pytest session scope to ensure that the build container and the
test image are only build once and then shared accross the
various individual tests.