bib: build to temporary directory and copy out
osbuild uses `cp -a` to export artifacts from a build and files are
(usually) owned by root.  The cp command tries to set the ownership to
root on the mounted volume from the host but if it's running in a podman
machine, this will fail.

Set the output directory to a temporary one inside the container and
copy the file to the /output directory when finished.