bib: check if we can chown in the output path
The canChownInPath() function creates a temporary file in the output
directory and tries to 'chown' it to the uid and gid of the caller.
This lets us know if osbuild will be able to preserve the permissions of
the output file when copying the results, which should be created with
the uid and gid of the caller of osbuild, to the output directory.

This is used to set export option when calling osbuild,
OSBUILD_EXPORT_FORCE_NO_PRESERVE_OWNER.  The option prevents osbuild
from trying to set the ownership of the files in the destination and
failing.