test: add image type and metadata to ImageBuildResult
Add the image type name and a metadata dict to the ImageBuildResult.
We will use the image type to decide what we can do with the build
result (like where and how to boot it).
The metadata can be used to communicate information to downstream tests.
Here we use it to include the AMI ID in the result.  We could have let
the tests that use the ImageBuildResult parse the log for the AMI ID,
but it's probably better if we only have to do it once and also, parsing
it in the image_type_fixture() lets us register the finalizer that will
deregister the AMI from EC2.

Parsing the log for the AMI ID is a bit dirty and is certainly not the
most robust way to get it.  We could later print it in a more
machine-readable format.