test: add test_ami_boots_in_aws()
A variant of the test_image_boots() test that only runs for the AMI
image type.
The function uses the AWS VM class to boot an AMI (given the AMI ID),
run a couple of commands over SSH, then tears it down.

As opposed to the local boot test, this one also runs on macOS.