README: De-emphasize specifying user configuration
Per the changes here, it is very much expected that users
create their own derived containers as the starting point
to the system, *not* just running the base image.

And once one does that, per the docs there are multiple solutions
for injecting and managing authentication credentials, lifecycled
with the *container* image as is often desired.

There's obviously use cases for injecting user configuration,
this change just de-emphasizes it and links to the existing
centos-bootc docs.