tmt: ask for bigger disks and get "df -h" output after testing
The tests are currently running out of disk space inside the
testing-farm environment. This commit should help debug this
by adding `df -h` output after the test is run.

It also asks for 120GB disk space (which is way more than we need
right now). Note that the docs say that the disk size is 250GB
by default for the RH ranch but we probably only get the 50G from
the public ranch by default because we are (currently) using
fedora39 as the test environment.

[1] https://docs.testing-farm.io/Testing%20Farm/0.1/test-request.html#hardware