bib: use the Linux runner for building disk images
Distro-specific runners try to set up a build environment to satisfy
certain expectations for stages (things they might expect from a live OS
environment).  One of the things the Fedora runner does is run
`systemd-tmpfiles --create`.  This causes a lot of warnings in our new
buildroot, like the following:

    /usr/lib/tmpfiles.d/rpm-ostree-autovar.conf:31: Failed to resolve user 'rpcuser': No such process

likely because we are running in an "ostree-based" tree that isn't fully
deployed (/etc/passwd only contains root).

The Linux runner is a fallback runner that doesn't perform any
preparations.  This should be adequate for the stages we use in bib.