Add selinux-policy-targeted and distribution-gpg-keys
This is part of reusing the bib container to build base images;
rpm-ostree wants to have *a* policy to bootstrap from currently.

For distribution-gpg-keys, it allows `.repo` files to reference
those unconditionally.

Also, reuse a "make package lists declarative" model I've used
in other places.  It'd be nice to try to semi-standardize this
of course, or even better upstream it into dnf.

We're looking at adding a different Containerfile in
https://github.com/osbuild/bootc-image-builder/pull/174
and this way it can source the same requirements.