bib: port to `bootc install to-filesystem` version of images
With the latest version of images we now use the
`bootc install to-filesystem` feature instead of our own ostree
deploy pipeline. This means that most of our customizations need
to be disabled and we need to decide if we will port them in some
form or update our README to encourage the use of derrived containters
for them.