bib: include fixed qemu for bib#207 and support using it
This commit allows working around the issue described in
https://github.com/osbuild/bootc-image-builder/issues/207

The workaround is that the container will register it's own
binfmt-misc handler with the fixed `qemu-user` that got
submitted in
https://lists.nongnu.org/archive/html/qemu-devel/2024-02/msg03971.html
when the environment `BIB_FORCE_INTERNAL_QEMU=1` is set.

As this modifies the global state of the machine it's not something
we want to do but might be a way to unblock people who need this
functionatlity.