workflow: make `pytest` less verbose
This is a bit of an experiment - we currently run `pytest` with
`--capture=no` which means we get realtime updates as the tests
run in GH actions. This is nice but it also leads to extremly
verbose test output which is hard to read because on failure
all test output is visible not just the failed one(s).

So this switches to the default mode of pytest to catpure the
output of successful tests and only show the output of tests
that failed. We lose the ability to view tests in realtime
but that is (nowdays) rarely needed (plus it's trivial to
run the tests locally).