Vendor Go dependencies
I'm hitting the pain where I want to hack on osbuild/images,
and run in a container.
We have the devel containerfile, but it requires some mental
energy to use.  Also, for the case where I don't also need to
hack on the osbuild binary it's a bit overkill.

Vendoring our Go dependencies makes the *main* Containerfile
immediately useful for local builds, I can `go mod replace`
with a local dir (just remembering to `go mod vendor`) or
even if I just want to truly hack something up I can edit
the `vendor/` dir directly.

We don't have to merge this but...I may just keep this
around and reuse it and reapply in situations like this.
Unsure.

Signed-off-by: Colin Walters <walters@verbum.org>