test: stop caching ami artifacts for now
The "ami" image from the cache is reused regardless of if it is
a local boot test or a boot test in AWS. This is a problem because
when the local image is generated first (without upload credentials
) it will never have been uploaded to AWS but the cache still tries
to reuse it. Worse because the filename of raw and ami are identical
it will reuse a raw image.

For now to unblock us skip AMI caching. This should be fixed in a
followup and this commit contains a todo with various options.