test: change build_images() to not assert because order is random
The order of the test runs is random but the build_images helper
makes assumptions that do not match this. I.e. it has an assert
that assumes that because all images are build the cache will
contain either all images requested or none. However if the
sequence of events is that
- build single qcow2 in test_image_boots
- build multi-output test
at this point the cache will have a qcow2 image already and
the assertion is violated and the test dies even though that
is a okay scenario.

This commit remove the assertion.

The helper grew a bit too big, we should rework it and also add
tests for it (recursion ftw!). Alternatively we could use something
like pytest-order to ensure that the if the multi test runs it
runs first.