build: Add .dockerignore and a comment
It's best practice to trim out stuff we don't need in the
container build from the context directory; this way if I e.g.
edit `tests` it won't invalidate the build caches.

Similarly, it's a really good idea to exclude things like `bin/`
from the builds - you don't want to accidentally somehow be
using incorrect cached *final* binaries.

But `Containerfile.hack` is all about bypassing that, so add
a comment for why we use the bind mount.