bib: add --filesystems commandline option
This commit adds support for basic filesystem customization. The
main use-case right now is to allow to create a bigger default
rootfs. This can be done via:
```
$ bootc-image-builder --filesystems "/:20G"
```
This is conceptually very similar to PR#124 but instead of blueprints
it is purely commandline driven. The rational is that it's unclear
what customization format we eventually want to support but it is
clear that we need a way to drive this from the commandline.

The format is designed to be extensible to a certain extend, so
```
bib --filesystems "/:20G,/var:100G:xfs"
```
is currently a strawman for supporting more customizations in
the future.

This PR is limited to "/" because we currently have no way to
inject /etc/fstab or systemd-mount units (the bootc --copy-etc
PR will probably helper here in the future).