bib: connect podman stdout to stderr when doing the initial podman pull
This should help with the potential UX issue that a podman pull
can take a while and for the user it looks like it's hanging.

It is a bit of a hack, we cannot use "stdout" when the "manifest"
subcommand is used as stdout must be pure json then. So it's this
or we just move the code into manifestFromCobra and only connect
stdout if `cmd.Use != "manifest"`. This is simpler.