bib: add -experimental-root-ssh-authorized-key option
With the move to bootc install to-filesystem we cannot currently
customize users except for roots ssh authorized_keys. So make
this the only allowed option. Note that it would be nicer if that
would be symetric to bootc --root-ssh-authorized-keys that takes
a file but our image customizations currently only support a single
key. So start here by marking it experimental to unblock tests and
fix images to eventually be symetric.