README.md: remove all traces of -config, update example
This commit removes the -config options from bib. Blueprint customizations
are no longer supported with the move to bootc install to-filesystem.
This is fine but we need to reflect it in our docs.

It also updates the example with an example how to create an image
with an user. This might be slightly controversial but I think it's
rather important to give users a quick path to success (and the
updated README also links to the upstream docs for a more indepth
discussion).