test: add `arch` argument to `make_container()` test helper
This is useful for cross arch container buildtesting. Note that
this always pass `--arch` when doing `podman build` because
without that the default behavior is to pull whatever arch was
pulled for this image ref last but we want "native" if nothing
else is specified.

Note: This just passes `platform.uname().machine` which uses the
kernel architecture names. podman seems to translate those kernel
arch names to go arches automatically.