bib: workaround missing BLKBSZSET in cross arch build
During a cross arch build we need to be conservative, i.e.  we can
only use the default ext4 because if xfs is select we run into the
issue that mkfs.xfs calls "ioctl(BLKBSZSET)" which is missing in
qemu-user, once the fix in
https://www.mail-archive.com/qemu-devel@nongnu.org/msg1037409.html
is merged we should revert this commit.