README.md: target c9s because ELN is more aggressive and has bugs
A while ago there was PR#17 that emphasised centos stream9 instead
of fedora:eln. We could not take it back then because the iso
installer was not quite ready for centos and there were hardcoded
EFI names. But now everything is in place so this cherry picks
the original changes from Colin.

See also https://github.com/osbuild/bootc-image-builder/pull/17