tmt: increase test duration to 3h
With the addition of the iso tests the testsuite runs slower now.
We need to think about if we really want to run all our tests
in testingfarm, the machien we get there is a bit on the slow
side and all we really need from it is the SELinux tests that
we cannot run on GH actions. But for now just increase the time
so that the tests pass again.