bib: deprecate the `--local` flag
Pulling container images before building would break in the case of
authenticated images on podman machine, since the auth file lives on the
host and not podman machine and won't know about it.

This commit silently deprecates the `--local` flag, meaning the flag can
still be passed by the user, without any effects and without warning the
user. This, however, does put the requirement on the user to ensure that
the container is in local storage before initiating the build.