bib: add `pull=missing` flag and logic
Add the `pull=missing` flag and set it as the default. Add the logic to
check if the container is missing from local storage and pull the image
if it is not there.

Note: it is a little difficult to check for cross-arch images if a
container already exists for the host's architecture. If the image
exists and the container arch doesn't match, just pull the image.