bib: disable `--tls-verify` flag
Since all containers are coming from local storage and require the user
to pull in the container before-hand, we can disable the `--tls-verify`
flag. The containers will not be resolved from a remote registry but
rather from the local container store. As with the removal of the
`--local` flag, we silently deprecate the flag - the flag will not
having any effects and will not raise any errors or warnings either.