docs: add new HACKING.md
As pointed out in https://github.com/osbuild/bootc-image-builder/pull/479
we are currently lacking a HACKING.md or similar to explain how
to run our tests. As this is non-obvious this commit adds a new
`HACKING.md` with the information.

It's a bit bare-bone and it would also be nice to create something
like a `get-test-deps.sh` or similar helper that is then used in our
various GH/tmt actions and can also be used by the user to get
up-to-date test dependencies. But that is something for a followup :)