README: add installer customizations
Add an example for installer customizations to the README.  Include the
same example in both toml and json format for better readability.
Unlike other examples, the toml format is shown first since it's easier
to work with.

Add a note for users explaining that the ostreecontainer command is
automatically added to the kickstart file when building an ISO.  Mention
that no other command is added so the user should include any other
options they need.

Link to the docs for up-to-date information on interactions with other
commands.