test: rework how test cases are represented/generated
The old way of generating test cases was via strings that would
use a `,` to differenciate between various parameters like
container_ref or image_type.

With the adding of the rootfs this becomes unwieldy (arguable it
was unwieldy already and now it's unentable). This commit introduces
a proper "TestCase" class that makes generating and reading the
test cases a lot eaiser (IMHO).

Note that this was prompted because of PR#439 where we want to
make the rootfs tests on fedora "btrfs".