Install osbuild-depsolve-dnf in container using dnf
The depsolver should not be installed by copying it into the container,
it should be installed from its package. Eventually it will not be a
single script (see https://github.com/osbuild/osbuild/pull/1776) and
will depend on dnf5 instead of dnf.

This adds an InstallPackages function to the container which uses podman
and dnf to install a list of package names and uses it to install the
osbuild-depsolve-dnf package. The execution path is also adjusted to be
the installed location.

Resolves: RHEL-50632