container: Add a test for InstallPackages
and fix the Stop function for podman versions that do remove the image
when stopped. When stopping the image if the removal fails because it
doesn't exist, don't return an error.

Add a test for container.InstallPackages that checks to make sure
osbuild-depsolve-dnf can be installed and the expected file is at
/usr/libexec/osbuild-depsolve-dnf

Related: RHEL-50632