bib: switch arch64 back from "gpt" to "dos" to unbreak PIs
The raspberry pi3 and newer version with outdated firmware
apparently do not support "gpt" partition tables. This
commit moves back to "dos" for maximum compatibility.

Once we have support to ship default partition tables
inside the container we can drop this again but for now
it seems the most common users of bootc on aarch64 are some
variant of a RPi so ideally the experience is as smooth as
possible there.