test: verify filesystem sizes when boot testing
During the boot test (test_image_boots()), run 'df' in the VM and verify
that the partitions and filesystems on disk match the filesystem
customizations from the build config.

Previously, we only had a check for /sysroot, but that wasn't doing
anything because there was no call to 'df' preceding it.  The current
test first checks that every mountpoint specified in the build config
exists in the VM and that the minsizes are respected (with some offsets
that will be investigated separately).