bib: propagate error messages from policy.Check()
The errors returned by policy.Check() can contain a reason for why a
path was rejected, especially when the rejection is not based on the
path policy itself.  For example, a non-absolute or non-canonical path
will be rejected.  This information is useful for the user, so let's not
discard that extra information and return all error messages we
encountered.

Co-authored-by: Michael Vogt <michael.vogt@gmail.com>